feat(*): Adds graceful deletion support
This partially implements the modify logic to handle graceful deletion.
Please note that some of this is not the most grateful and should
probably be refactored down the line, but it is functional.

I also found a small bug where we were reporting an error when creating
volumes if they already existed and fixed that.

Closes #154