ref(*): Removes use of deprecated `Informer` from the `kube` crate (#396)
* ref(*): Removes use of deprecated `Informer` from the `kube` crate

This replaces all uses of the `Informer` with the new `watcher` type
from the `kube_runtime` crate. I also fixed a bunch of clippy lints I
found along the way

* Update features to include new runtime crate

* ref(*): Handles watch restarts properly

Now if there is a restart event, we will resync all changes.

This also contains a refactor around pod keys by introducing the new
`PodKey` type. This serves 2 main purposes. The first is it makes it
clear that we are using key for a pod rather than it just being a
`String`. The second is that instead of it being an arbitrary string,
it now has separate fields that define the key. Arbitrary strings and
separators can come back to bite, so this makes it a concrete type
with separate fields

* fix(ci): Increases delay for namespace wait

* Fixes race condition in test where the next process thought a
CSR already existed