chore(*): Bump the krator and derive versions
This is in preparation for the 0.4 release of Krator. This is needed
because supporting k8s 1.21 required a version bump of the k8s_openapi
crate, which contained breaking changes. Because of those breaking
changes, it is a breaking change as well for krator. This still supports
any version of k8s that k8s_openapi does

Signed-off-by: Taylor Thomas <taylor.thomas@microsoft.com>